Refactor to allow Murfey servers to operate for any instrument in configuration (#352)
Also allows one server to use another for auth if configured. This will allow the use of only a single central server for auth

+++ src/murfey/server/websocket.py
@@ -4,13 +4,14 @@ import asyncio
 import json
 import logging
 from datetime import datetime
-from typing import Any, Dict, Generic, TypeVar
+from typing import Any, Dict, Generic, TypeVar, Union
 
 from fastapi import APIRouter, WebSocket, WebSocketDisconnect
 from sqlmodel import select
 
 import murfey.server.prometheus as prom
 from murfey.server.murfey_db import get_murfey_db_session
+from murfey.util import sanitise
 from murfey.util.db import ClientEnvironment
 from murfey.util.state import State, global_state
 
@@ -22,14 +23,21 @@ log = logging.getLogger("murfey.server.websocket")
 
 class ConnectionManager(Generic[T]):
     def __init__(self, state: State[T]):
-        self.active_connections: Dict[int, WebSocket] = {}
+        self.active_connections: Dict[int | str, WebSocket] = {}
         self._state = state
         self._state.subscribe(self._broadcast_state_update)
 
-    async def connect(self, websocket: WebSocket, client_id: int):
+    async def connect(
+        self, websocket: WebSocket, client_id: int | str, register_client: bool = True
+    ):
         await websocket.accept()
         self.active_connections[client_id] = websocket
-        self._register_new_client(client_id)
+        if register_client:
+            if not isinstance(client_id, int):
+                raise ValueError(
+                    "To register a client the client ID must be an integer"
+                )
+            self._register_new_client(client_id)
         await websocket.send_json({"message": "state-full", "state": self._state.data})
 
     @staticmethod
@@ -40,15 +48,20 @@ class ConnectionManager(Generic[T]):
         murfey_db.commit()
         murfey_db.close()
 
-    def disconnect(self, websocket: WebSocket, client_id: int):
+    def disconnect(
+        self, websocket: WebSocket, client_id: int | str, unregister_client: bool = True
+    ):
         self.active_connections.pop(client_id)
-        murfey_db = next(get_murfey_db_session())
-        client_env = murfey_db.exec(
-            select(ClientEnvironment).where(ClientEnvironment.client_id == client_id)
-        ).one()
-        murfey_db.delete(client_env)
-        murfey_db.commit()
-        murfey_db.close()
+        if unregister_client:
+            murfey_db = next(get_murfey_db_session())
+            client_env = murfey_db.exec(
+                select(ClientEnvironment).where(
+                    ClientEnvironment.client_id == client_id
+                )
+            ).one()
+            murfey_db.delete(client_env)
+            murfey_db.commit()
+            murfey_db.close()
 
     async def broadcast(self, message: str):
         for connection in self.active_connections:
@@ -63,11 +76,13 @@ class ConnectionManager(Generic[T]):
             )
 
     async def set_state(self, attribute: str, value: T):
-        log.info(f"State attribute {attribute!r} set to {value!r}")
+        log.info(
+            f"State attribute {sanitise(attribute)!r} set to {sanitise(str(value))!r}"
+        )
         await self._state.set(attribute, value)
 
     async def delete_state(self, attribute: str):
-        log.info(f"State attribute {attribute!r} removed")
+        log.info(f"State attribute {sanitise(attribute)!r} removed")
         await self._state.delete(attribute)
 
 
@@ -90,13 +105,40 @@ async def websocket_endpoint(websocket: WebSocket, client_id: int):
             except Exception:
                 await manager.broadcast(f"Client #{client_id} sent message {data}")
     except WebSocketDisconnect:
-        log.info(f"Disconnecting Client {client_id}")
+        log.info(f"Disconnecting Client {int(sanitise(str(client_id)))}")
         murfey_db = next(get_murfey_db_session())
         client_env = murfey_db.exec(
             select(ClientEnvironment).where(ClientEnvironment.client_id == client_id)
         ).one()
         prom.monitoring_switch.labels(visit=client_env.visit).set(0)
         manager.disconnect(websocket, client_id)
+        await manager.broadcast(f"Client #{client_id} disconnected")
+        await manager.delete_state(f"Client {client_id}")
+
+
+@ws.websocket("/connect/{client_id}")
+async def websocket_connection_endpoint(
+    websocket: WebSocket, client_id: Union[int, str]
+):
+    await manager.connect(websocket, client_id, register_client=False)
+    await manager.broadcast(f"Client {client_id} joined")
+    await manager.set_state(f"Client {client_id}", "joined")
+    try:
+        while True:
+            data = await websocket.receive_text()
+            try:
+                json_data = json.loads(data)
+                if json_data.get("type") == "log":  # and isinstance(json_data, dict)
+                    json_data.pop("type")
+                    await forward_log(json_data, websocket)
+                elif json_data.get("message") == "refresh":
+                    await manager.broadcast(json.dumps(json_data))
+
+            except Exception:
+                await manager.broadcast(f"Client #{client_id} sent message {data}")
+    except WebSocketDisconnect:
+        log.info(f"Disconnecting Client {int(sanitise(str(client_id)))}")
+        manager.disconnect(websocket, client_id, unregister_client=False)
         await manager.broadcast(f"Client #{client_id} disconnected")
         await manager.delete_state(f"Client {client_id}")
 
@@ -138,4 +180,12 @@ async def close_ws_connection(client_id: int):
     log.info(f"Disconnecting {client_id_str}")
     manager.disconnect(manager.active_connections[client_id], client_id)
     prom.monitoring_switch.labels(visit=visit_name).set(0)
+    await manager.broadcast(f"Client #{client_id} disconnected")
+
+
+@ws.delete("/connect/{client_id}")
+async def close_unrecorded_ws_connection(client_id: Union[int, str]):
+    client_id_str = str(client_id).replace("\r\n", "").replace("\n", "")
+    log.info(f"Disconnecting {client_id_str}")
+    manager.disconnect(manager.active_connections[client_id], client_id)
     await manager.broadcast(f"Client #{client_id} disconnected")
